requirements.txt
pytest
pylint
ezdxf
pyyaml
